requirements.txt
caldav
icalendar
pygame
pillow
pyowm
pytz
